// src/api/fbs/mybroker/api-payload.ts
// automatically generated by the FlatBuffers compiler, do not modify

 

import { AuthOk } from '../mybroker/auth-ok';
import { ChatDetail } from '../mybroker/chat-detail';
import { ChatRowList } from '../mybroker/chat-row-list';
import { ChatsAndFavs } from '../mybroker/chats-and-favs';
import { Empty } from '../mybroker/empty';
import { NestedPostT } from '../mybroker/nested-post-t';
import { PostList } from '../mybroker/post-list';
import { PostLocationList } from '../mybroker/post-location-list';
import { PostPage } from '../mybroker/post-page';
import { UsersList } from '../mybroker/users-list';


export enum ApiPayload {
  NONE = 0,
  Empty = 1,
  PostPage = 2,
  PostList = 3,
  NestedPostT = 4,
  PostLocationList = 5,
  ChatRowList = 6,
  ChatDetail = 7,
  ChatsAndFavs = 8,
  AuthOk = 9,
  UsersList = 10
}

export function unionToApiPayload(
  type: ApiPayload,
  accessor: (obj:AuthOk|ChatDetail|ChatRowList|ChatsAndFavs|Empty|NestedPostT|PostList|PostLocationList|PostPage|UsersList) => AuthOk|ChatDetail|ChatRowList|ChatsAndFavs|Empty|NestedPostT|PostList|PostLocationList|PostPage|UsersList|null
): AuthOk|ChatDetail|ChatRowList|ChatsAndFavs|Empty|NestedPostT|PostList|PostLocationList|PostPage|UsersList|null {
  switch(ApiPayload[type]) {
    case 'NONE': return null; 
    case 'Empty': return accessor(new Empty())! as Empty;
    case 'PostPage': return accessor(new PostPage())! as PostPage;
    case 'PostList': return accessor(new PostList())! as PostList;
    case 'NestedPostT': return accessor(new NestedPostT())! as NestedPostT;
    case 'PostLocationList': return accessor(new PostLocationList())! as PostLocationList;
    case 'ChatRowList': return accessor(new ChatRowList())! as ChatRowList;
    case 'ChatDetail': return accessor(new ChatDetail())! as ChatDetail;
    case 'ChatsAndFavs': return accessor(new ChatsAndFavs())! as ChatsAndFavs;
    case 'AuthOk': return accessor(new AuthOk())! as AuthOk;
    case 'UsersList': return accessor(new UsersList())! as UsersList;
    default: return null;
  }
}

export function unionListToApiPayload(
  type: ApiPayload, 
  accessor: (index: number, obj:AuthOk|ChatDetail|ChatRowList|ChatsAndFavs|Empty|NestedPostT|PostList|PostLocationList|PostPage|UsersList) => AuthOk|ChatDetail|ChatRowList|ChatsAndFavs|Empty|NestedPostT|PostList|PostLocationList|PostPage|UsersList|null, 
  index: number
): AuthOk|ChatDetail|ChatRowList|ChatsAndFavs|Empty|NestedPostT|PostList|PostLocationList|PostPage|UsersList|null {
  switch(ApiPayload[type]) {
    case 'NONE': return null; 
    case 'Empty': return accessor(index, new Empty())! as Empty;
    case 'PostPage': return accessor(index, new PostPage())! as PostPage;
    case 'PostList': return accessor(index, new PostList())! as PostList;
    case 'NestedPostT': return accessor(index, new NestedPostT())! as NestedPostT;
    case 'PostLocationList': return accessor(index, new PostLocationList())! as PostLocationList;
    case 'ChatRowList': return accessor(index, new ChatRowList())! as ChatRowList;
    case 'ChatDetail': return accessor(index, new ChatDetail())! as ChatDetail;
    case 'ChatsAndFavs': return accessor(index, new ChatsAndFavs())! as ChatsAndFavs;
    case 'AuthOk': return accessor(index, new AuthOk())! as AuthOk;
    case 'UsersList': return accessor(index, new UsersList())! as UsersList;
    default: return null;
  }
}